requirements.txt
asgiref==3.9.2
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
djangorestframework==3.16.1
dnspython==2.8.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
mongoengine==0.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pymongo==3.11.4
pyOpenSSL==25.3.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
zope.interface==8.0.1
